requirements.txt
arabic-reshaper==3.0.0
asgiref==3.8.1
asn1crypto==1.5.1
beautifulsoup4==4.10.0
Brotli==1.1.0
certifi==2025.1.31
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.4.1
click==8.1.8
contourpy==1.3.1
cryptography==44.0.2
cssselect2==0.8.0
cycler==0.12.1
defusedxml==0.7.1
distlib==0.3.9
distro==1.7.0
Django==5.2
django-allauth==65.7.0
django-filter==25.1
djangorestframework==3.16.0
drf-yasg==1.21.10
filelock==3.18.0
fonttools==4.57.0
fpdf==1.7.2
fpdf2==2.8.2
greenlet==3.1.1
html5lib==1.1
idna==3.10
inflection==0.5.1
kiwisolver==1.4.8
lxml==5.3.2
matplotlib==3.10.1
numpy==2.2.4
oscrypto==1.3.0
packaging==24.2
pandas==2.2.3
pillow==11.1.0
pipenv==2024.4.1
platformdirs==4.3.7
playwright==1.51.0
psycopg2-binary==2.9.10
pycparser==2.22
pydyf==0.11.0
pyee==12.1.1
pyHanko==0.26.0
pyhanko-certvalidator==0.26.8
pyparsing==3.2.3
pypdf==5.4.0
pyphen==0.17.2
python-bidi==0.6.6
python-dateutil==2.9.0.post0
python-docx==0.8.11
pytz==2025.2
PyYAML==6.0.2
qrcode==8.1
reportlab==4.3.1
requests==2.32.3
six==1.17.0
soupsieve==2.3.1
sqlparse==0.5.3
ssh-import-id==5.11
supervisor==4.2.1
svglib==1.5.1
tinycss2==1.4.0
tinyhtml5==2.0.0
typing_extensions==4.13.1
tzdata==2025.2
tzlocal==5.3.1
uritemplate==4.1.1
uritools==4.0.3
urllib3==2.3.0
virtualenv==20.30.0
weasyprint==65.0
webencodings==0.5.1
xhtml2pdf==0.2.17
zopfli==0.2.3.post1
